requirements.txt
Flask==2.2.2
Pillow==9.4.0
rembg==2.0.30
